separate files for space and quote handling
