+++ input_validation.c
@@ -13,42 +13,6 @@
 #include "minishell.h"
 #include "builtins.h"
 
-int	skip_quotes(char *line, int i)
-{
-	char	quote;
-
-	quote = line[i];
-	i++;
-	while (line[i] && line[i] != quote)
-		i++;
-	return (i);
-}
-bool	has_unclosed_quotes(char *line)
-{
-	int	count;
-	int	i;
-
-	i = 0;
-	count = 0;
-	while (line[i])
-	{
-		if (line[i] == '\'' || line[i] == '\"')
-			count++;
-		i++;
-	}
-	if (count % 2 == 0)
-		return (false);
-	return (true);
-}
-
-bool	check_space(char ch)
-{
-	if (ch == ' ' || ch == '\t' || ch == '\n' || ch == '\r' || \
-		ch == '\f' || ch == '\v')
-		return (true);
-	return (false);
-}
-
 int	input_validation_pipes(char *input)
 {
 	int	i;
@@ -110,25 +74,4 @@ void	error_message(char *msg)
 	//ft_putstr_fd("Error\n", 2);
 	ft_putendl_fd(msg, 2);
 	exit (EXIT_FAILURE);
-}
-
-void	put_one_space(char *line)
-{
-	int i;
-
-	i =  0;
-	while (line[i] && check_space(line[i]))
-			i++;
-	while (line[i])
-	{
-		while (line[i] && !check_space(line[i]))
-		{
-			printf("%c", line[i]);
-			i++;
-		}
-		while (line[i] && check_space(line[i]))
-			i++;
-		if (line[i] && !check_space(line[i]))
-		printf(" ");
-	}
 }
