+++ quotes_handle.c
@@ -1,0 +1,91 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   quotes_handle.c                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: epolkhov <epolkhov@student.42.fr>          #+#  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024-06-26 10:45:47 by epolkhov          #+#    #+#             */
+/*   Updated: 2024-06-26 10:45:47 by epolkhov         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+#include "builtins.h"
+
+int	skip_quotes(char *line, int i)
+{
+	char	quote;
+
+	quote = line[i];
+	i++;
+	while (line[i] && line[i] != quote)
+		i++;
+	return (i);
+}
+
+bool	has_unclosed_quotes(char *line)
+{
+	int	count;
+	int	i;
+
+	i = 0;
+	count = 0;
+	while (line[i])
+	{
+		if (line[i] == '\'' || line[i] == '\"')
+			count++;
+		i++;
+	}
+	if (count % 2 == 0)
+		return (false);
+	return (true);
+}
+
+void	remove_quotes(char *str)
+{
+	int	i;
+	int	j;
+
+	i = 0;
+	j = 0;
+	if (!str)
+		return ;
+	while (str[i])
+	{
+		if (str[i] != '\"' && str[i] != '\'')
+		{
+			str[j] = str[i];
+			j++;
+		}
+		i++;
+	}
+	str[j] = '\0';
+}
+
+int	check_input_quotes_pipe(char *line)
+{
+	int	i;
+	int	in_quote;
+
+	i = 0;
+	in_quote = 0;
+	while (line[i])
+	{
+		if (line[i] == '\"' || line[i] == '\'')
+		{
+			in_quote = !in_quote;
+			i++;
+			continue ;
+		}
+		if (line[i] == '|' && !in_quote)
+			line[i] = 31;
+		i++;
+	}
+	if ((in_quote || has_unclosed_quotes(line)))
+	{
+		ft_putendl_fd("Syntax error: unclosed quotes", 2);
+		return (1);
+	}
+	return (0);
+}
